Consolidate into MessageReplyForm; adding/viewing attachments functional in full layout; #8
