+++ client/app/messages/Attachment.jsx
@@ -3,15 +3,17 @@
 var _ = require('underscore');
 var React = require('react');
 var truncate = require('../utils/truncate');
+var config = require('../utils/config');
 
 var Attachment = React.createClass({
 
   render: function() {
+    console.log('render attachment', config.apiUrl)
     var att = this.props.attachment;
     var filename = truncate(att.filename, 23);
     var thumbnail = att.thumbnail;
     var fileType;
-    var downloadUrl = 'https://gingerhq.com/' + att.attachment;
+    var downloadUrl = config.apiUrl + att.attachment;
     // get a thumbnail from the file extension
     if (!thumbnail) {
       fileType = filename.split('.')[1];
@@ -20,11 +22,11 @@ var Attachment = React.createClass({
       }
       thumbnail = '/assets/img/file-icons/' + fileType + '.gif';
     }
-    thumbnail = 'https://gingerhq.com' + thumbnail;
+    thumbnail = config.apiUrl + thumbnail;
     return (
       <li className="attachment-item">
         <a href={downloadUrl}>
-          <img className="item-icon" src={thumbnail} alt={filename} /> 
+          <img className="item-icon" src={thumbnail} alt={filename} />
           <span className="item-name">{filename} </span>
         </a>
       </li>
