+++ client/app/messages/Message.jsx
@@ -16,8 +16,6 @@ var MessageContent = require('./MessageContent.jsx');
 var MessageReplyForm = require('./MessageReplyForm.jsx');
 var VotesListView = require('./VotesList.jsx');
 var LayoutStore = require('../layout/LayoutStore');
-var cookies = require('../utils/cookies');
-var config = require('../utils/config');
 
 var MessageView = React.createClass({
   mixins: [eventsMixin],
@@ -122,34 +120,6 @@ var MessageView = React.createClass({
     this.emitter.on('message:focus', this.handleFocusChange);
   },
 
-  handleFileInput: function (event) {
-    console.log('file input',event.target.files)
-
-    //POST file
-    var uploadIframe = window.frames.uploadFileIframe,
-        data = new FormData();
-
-    uploadIframe.xhr = new XMLHttpRequest();
-    var xhr = uploadIframe.xhr;
-
-    var message = store.find('messages', this.props.message.url);
-    data.append('attachment', event.target.files[0]);
-
-    //xhr.upload.addEventListener('progress', this.uploadProgress, false);
-    xhr.addEventListener('load', this.uploadSuccess, false);
-
-    //TODO get API URL
-    var uploadUrl = config.apiUrl + urls.get('api:fileUpload') + '?format=json';
-    xhr.open('POST', uploadUrl, true);
-    xhr.setRequestHeader('X-CSRFToken', cookies.getItem('csrftoken'));
-    xhr.setRequestHeader('Authorization', 'Token ' + localStorage.getItem('apiKey'));
-    xhr.send(data);
-  },
-
-  uploadSuccess: function () {
-    console.log('success!')
-  },
-
   render: function() {
     // shortcuts
     log.debug('render', this.state)
@@ -215,16 +185,10 @@ var MessageView = React.createClass({
           </div>
         </div>
         {this.state.reply && (
-          <div>
-            <div>
-              <iframe id="uploadFileIframe" name="uploadFileIframe"><html><body></body></html></iframe>
-              <input type='file' onChange={this.handleFileInput}/>
-            </div>
             <MessageReplyForm
               messageId={message.id}
               parent_url={urls.get('api:messageChange', {'message_id': message.id})}
               callback={this.handleReplySuccess} />
-          </div>
         )}
       </div>
     );
